+++ packages/main/src/Title.ts
@@ -39,11 +39,7 @@ import titleCss from "./generated/themes/Title.css.js";
 class Title extends UI5Element {
 	/**
 	 * Defines how the text of a component will be displayed when there is not enough space.
-	 * Available options are:
-	 * <ul>
-	 * <li><code>None</code> - The text will be truncated with an ellipsis.</li>
-	 * <li><code>Normal</code> - The text will wrap. The words will not be broken based on hyphenation.</li>
-	 * </ul>
+	 * <br><b>Note:</b> for option "Normal" the text will wrap and the words will not be broken based on hyphenation.
 	 *
 	 * @name sap.ui.webc.main.Title.prototype.wrappingType
 	 * @type {sap.ui.webc.main.types.WrappingType}