+++ packages/playground/_stories/fiori/IllustratedMessage/IllustratedMessage.stories.ts
@@ -16,7 +16,7 @@ const component = "ui5-illustrated-message";
 
 export default {
     title: "Fiori/IllustratedMessage",
-    component,
+    component: "IllustratedMessage",
     parameters: {
         docs: {
             page: DocsPage({ ...componentInfo, component }),