+++ packages/playground/_stories/main/Carousel/Carousel.stories.ts
@@ -17,7 +17,7 @@ const component = "ui5-carousel";
 
 export default {
 	title: "Main/Carousel",
-	component,
+	component: "Carousel",
 	parameters: {
 		docs: {
 			page: DocsPage({ ...componentInfo, component })
@@ -42,14 +42,14 @@ const Template: UI5StoryArgs<Carousel, StoryArgsSlots> = (args) => {
 </ui5-carousel>`;
 }
 
-export const SingleItemPerPage = Template.bind({});
-SingleItemPerPage.args = {
+export const Basic = Template.bind({});
+Basic.args = {
 	cyclic: true,
 	default: `<img src="../assets/images/sample1.jpg" alt="Landscape 1">
 	<img src="../assets/images/sample2.jpg" alt="Landscape 2">
 	<img src="../assets/images/sample3.jpg" alt="Bulb">`
 };
-SingleItemPerPage.decorators = [
+Basic.decorators = [
 	(story) => html`<style>
 	img {
 		max-height: 30rem;